Extract user auth functions into separate service
