+++ src/auth/auth.controller.spec.ts
@@ -8,7 +8,12 @@ import { UserService } from '../user/user.service';
 import { UserToken } from '../user/entities/user-token.entity';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
-import { PasswordResetRequestCreatedDto, PasswordResetRequestDto, SetNewPasswordDto, TokenDto, } from './dto/auth';
+import {
+  PasswordResetRequestCreatedDto,
+  PasswordResetRequestDto,
+  SetNewPasswordDto,
+  TokenDto,
+} from './dto/auth';
 import { JwtModule } from '@nestjs/jwt';
 import * as httpMocks from 'node-mocks-http';
 import { NotificationServiceInterface } from '../notification/types/notification-service';
@@ -16,6 +21,7 @@ import { UserSession } from './entities/user-session.entity';
 import { UserIdentifier } from './types/auth';
 import { ConfigService } from '@nestjs/config';
 import { NotFoundException } from '@nestjs/common';
+import { UserAuthService } from '../user/user-auth.service';
 
 const defaultToken = 'insecure-jwt-token-used-for-testing-only';
 
@@ -28,6 +34,7 @@ describe('AuthController', () => {
   let authController: AuthController;
   let authService: AuthService;
   let userService: UserService;
+  let userAuthService: UserAuthService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
@@ -41,6 +48,7 @@ describe('AuthController', () => {
       providers: [
         AuthService,
         UserService,
+        UserAuthService,
         CryptoService,
         {
           provide: ConfigService,
@@ -72,6 +80,7 @@ describe('AuthController', () => {
 
     authController = module.get<AuthController>(AuthController);
     userService = module.get<UserService>(UserService);
+    userAuthService = module.get<UserAuthService>(UserAuthService);
     authService = module.get<AuthService>(AuthService);
   });
 
@@ -135,7 +144,7 @@ describe('AuthController', () => {
         token: 'y',
       } as ActivateUserDto;
       const spy = jest
-        .spyOn(userService, 'activateUser')
+        .spyOn(userAuthService, 'activateUser')
         .mockReturnValue(mockReturn);
 
       await authController.activateUser(mockUser);
@@ -196,7 +205,7 @@ describe('AuthController', () => {
         user: userMock,
       };
       const userServiceSpy = jest
-        .spyOn(userService, 'createPasswordResetTokenForUser')
+        .spyOn(userAuthService, 'createPasswordResetTokenForUser')
         .mockReturnValue(Promise.resolve(mockTokenResponse));
       const notificationServiceSpy = jest
         .spyOn(notificationServiceMock, 'sendPasswordResetRequestMessage')
@@ -210,7 +219,7 @@ describe('AuthController', () => {
 
     it('fails gracefully if the user is not found and does not send a message', async () => {
       jest
-        .spyOn(userService, 'createPasswordResetTokenForUser')
+        .spyOn(userAuthService, 'createPasswordResetTokenForUser')
         .mockImplementation(() => {
           throw new NotFoundException();
         });
@@ -224,7 +233,7 @@ describe('AuthController', () => {
 
     it('throws exceptions other than NotFoundError', async () => {
       jest
-        .spyOn(userService, 'createPasswordResetTokenForUser')
+        .spyOn(userAuthService, 'createPasswordResetTokenForUser')
         .mockImplementation(() => {
           throw new Error();
         });
@@ -240,7 +249,7 @@ describe('AuthController', () => {
   describe('passwordResetSet', () => {
     it('calls userService.resetPassword() with correct params', async () => {
       const serviceSpy = jest
-        .spyOn(userService, 'resetPassword')
+        .spyOn(userAuthService, 'resetPassword')
         .mockReturnValue(Promise.resolve());
       const setNewPasswordDto: SetNewPasswordDto = {
         password: 'x-x-x',
