+++ src/user/user.module.ts
@@ -5,11 +5,12 @@ import { UserService } from './user.service';
 import { UserController } from './user.controller';
 import { UtilsModule } from '../utils/utils.module';
 import { UserToken } from './entities/user-token.entity';
+import { UserAuthService } from './user-auth.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User, UserToken]), UtilsModule],
-  providers: [UserService],
+  providers: [UserService, UserAuthService],
   controllers: [UserController],
-  exports: [UserService],
+  exports: [UserService, UserAuthService],
 })
 export class UserModule {}
